Divided MainWindow.cs to manage MainWindow class easily
